requirements.txt
discord.py>=1.7.1
python-decouple
pillow
pyrebase
imageio
akinator.py
pygifsicle
fun-responses
discord-components
asyncio
anime-images-api
disputils
pet-pet-gif
